pyproject.toml
[tool.poetry]
name = "asdf"
version = "0.1.0"
description = "asdf"
authors = [

]
maintainers = [

]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.75.0"
uvicorn = { version = "^0.17.0", extras = ["standard"] }
pydantic = {version = "^1.9.0", extras = ["dotenv"]}
yarl = "^1.7.2"
ujson = "^5.1.0"
SQLAlchemy = {version = "^1.4", extras = ["mypy", "asyncio"]}
alembic = "^1.7.7"
asyncpg = {version = "^0.25.0", extras = ["sa"]}
redis = {version = "^4.3.3", extras = ["hiredis"]}
aiofiles = "^0.8.0"
httptools = "^0.3.0"
strawberry-graphql = { version = "^0.114.2", extras = ["fastapi"] }
psycopg2 = "^2.9.3"

[tool.poetry.dev-dependencies]
pytest = "^7.0"
flake8 = "^4.0.1"
mypy = "^0.961"
isort = "^5.10.1"
yesqa = "^1.3.0"
pre-commit = "^2.19.0"
wemake-python-styleguide = "^0.16.1"
black = "^22.3.0"
autoflake = "^1.4"
SQLAlchemy = {version = "^1.4", extras = ["mypy"]}
pytest-cov = "^3.0.0"
anyio = "^3.6.1"
pytest-env = "^0.6.2"
fakeredis = "^1.8.1"
httpx = "^0.22.0"
sqlacodegen = "3.0.0rc1"
add-trailing-comma = "^2.2.3"

[tool.isort]
profile = "black"
multi_line_output = 3
src_paths = ["asdf",]

[tool.mypy]
strict = true
ignore_missing_imports = true
allow_subclassing_any = true
allow_untyped_calls = true
pretty = true
show_error_codes = true
implicit_reexport = true
allow_untyped_decorators = true
warn_unused_ignores = false
warn_return_any = false
namespace_packages = true
plugins = ["sqlalchemy.ext.mypy.plugin"]

# Remove this and add `types-redis`
# when the issue https://github.com/python/typeshed/issues/8242 is resolved.
[[tool.mypy.overrides]]
module = [
    'redis.asyncio'
]
ignore_missing_imports = true

[tool.pytest.ini_options]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
    "ignore:.*unclosed.*:ResourceWarning",
]
env = [
    "ASDF_DB_BASE=asdf_test",
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
